# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FXREBATE PROJECT",
        "description": "REACT Progressive Web App / Stand on the shoulders of giants",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/companies/create": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Get form data for creating company",
                "operationId": "db75542fbf1e58735d2a645b79221378",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Create form endpoint"
                                        },
                                        "data": {
                                            "properties": {
                                                "brokers": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "zones": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "status_options": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
    },
...
}